Refactor flatten and unflatten to core package
