+++ package.json
@@ -22,6 +22,7 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "build": "rollup -c",
+    "build:watch": "rollup -c -w",
     "postinstall": "yarn build"
   },
   "repository": {
