+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
-export * from './types';
+export * from './interfaces';
 export * from './constants';
 export * from './utils';
 export * from './diff';
+export * from './steps';
