+++ packages/core/src/interfaces.ts
