+++ packages/editor/src/utils/collection.ts
@@ -1,169 +1,8 @@
 import * as F from 'editure-constants';
 import { Node, Element } from 'editure';
-import omit from 'lodash.omit';
-import { Article, Step, File, DiffBlock, Explain } from '@tuture/core';
+import { Article, Step } from '@tuture/core';
 
-import {
-  FILE,
-  STEP,
-  DIFF_BLOCK,
-  EXPLAIN,
-  STEP_END,
-  STEP_START,
-  FILE_START,
-  FILE_END,
-} from './constants';
 import { HeadingItem } from '../types';
-
-export function flatten(steps: Step[]): Node[] {
-  return steps.flatMap(({ commit, id, articleId, children }) => {
-    let stepExplainNumber = 0;
-
-    return [
-      {
-        type: STEP_START,
-        children: [{ text: '' }],
-      },
-      {
-        commit,
-        id,
-        articleId,
-        type: STEP,
-        children: [{ text: '' }],
-      },
-      ...children.flatMap((node: any) => {
-        if (node.type === EXPLAIN) {
-          stepExplainNumber = stepExplainNumber + 1;
-
-          if (stepExplainNumber === 1) {
-            return { ...node, flag: STEP_START };
-          } else if (stepExplainNumber === 2) {
-            return { ...node, flag: STEP_END };
-          }
-        }
-
-        if (node.type === FILE && node.display) {
-          const { file, display } = node;
-
-          let fileExplainNumber = 0;
-          const explainChildrenWithFlag = node.children.map((nodeItem: any) => {
-            if (nodeItem.type === EXPLAIN) {
-              fileExplainNumber = fileExplainNumber + 1;
-
-              if (fileExplainNumber === 1) {
-                return { ...nodeItem, flag: FILE_START };
-              } else if (fileExplainNumber === 2) {
-                return { ...nodeItem, flag: FILE_END };
-              }
-            }
-
-            return nodeItem;
-          });
-
-          return [
-            {
-              type: FILE_START,
-              children: [{ text: '' }],
-            },
-            {
-              file,
-              display,
-              commit,
-              type: FILE,
-              children: [{ text: '' }],
-            },
-            ...explainChildrenWithFlag,
-            {
-              type: FILE_END,
-              children: [{ text: '' }],
-            },
-          ];
-        }
-
-        return node;
-      }),
-      {
-        type: STEP_END,
-        children: [{ text: '' }],
-      },
-    ];
-  });
-}
-
-export function unflatten(fragment: Node[]) {
-  const steps = [];
-
-  let step: Partial<Step> = { children: [] };
-  let flag = '';
-  for (let i = 0; i < fragment.length; i++) {
-    const node = fragment[i];
-
-    switch (node.type) {
-      case STEP: {
-        step = { ...node, children: [] };
-        break;
-      }
-
-      case FILE: {
-        if (node.display) {
-          step.children!.push({ ...node, children: [] } as any);
-        } else {
-          step.children!.push(node as File);
-        }
-        break;
-      }
-
-      case DIFF_BLOCK: {
-        step.children!.slice(-1)[0].children.push(node as DiffBlock);
-
-        break;
-      }
-
-      case EXPLAIN: {
-        // In step: step_start || file_end
-        if (flag === 'step_start' || flag === 'file_end') {
-          step.children!.push(omit(node, 'flag') as Explain);
-        }
-
-        if (flag === 'file_start') {
-          step
-            .children!.slice(-1)[0]
-            .children.push(omit(node, 'flag') as Explain);
-        }
-
-        break;
-      }
-
-      case STEP_START: {
-        flag = 'step_start';
-        break;
-      }
-
-      case STEP_END: {
-        flag = 'step_end';
-        steps.push(step);
-
-        break;
-      }
-
-      case FILE_START: {
-        flag = 'file_start';
-        break;
-      }
-
-      case FILE_END: {
-        flag = 'file_end';
-        break;
-      }
-
-      default: {
-        step.children!.push(node as any);
-      }
-    }
-  }
-
-  return steps as Step[];
-}
 
 function isHeading(node: Node) {
   return [F.H1, F.H2, F.H3, F.H4, F.H5].includes(node.type);
